rearange files
