+++ frontEnd/src/components/Home/displayData.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./App.css";
+import "../../styles/App.css";
 
 const DisplayCrimeData = ({ crimeType, crimeNumber }) => {
   return (
