+++ frontEnd/src/components/Home/loading.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "./loading.css";
+import "../../styles/loading.css";
 
 const Spinner = () => {
   return <div className="loader"></div>;
