+++ frontEnd/src/components/Home/options.js
